📚 refactor: router migration
